+++ todoman/model.py
@@ -543,7 +543,7 @@ class Cache:
 
         return rv
 
-    def todos(self, all=False, lists=[], urgent=False, location='',
+    def todos(self, all=False, lists=[], priority=None, location='',
               category='', grep='', sort=[], reverse=True, due=None,
               complete=None, start=None):
         """
@@ -558,7 +558,6 @@ class Cache:
 
         :param bool all: If true, also return completed todos.
         :param list lists: Only return todos for these lists.
-        :param bool urgent: Only return urgent todos.
         :param str location: Only return todos with a location containing this
             string.
         :param str category: Only return todos with a category containing this
@@ -568,6 +567,8 @@ class Cache:
             with a ``-`` prepended will be used to sort in reverse order.
         :param bool reverse: Reverse the order of the todos after sorting.
         :param int due: Return only todos due within ``due`` hours.
+        :param str priority: Only return todos with priority at least as
+            high as specified.
         :param bool complete: If true, return completed tasks,
             else incomplete tasks
         :param start: Return only todos before/after ``start`` date
@@ -592,8 +593,9 @@ class Cache:
             q = ', '.join(['?'] * len(lists))
             extra_where.append('AND files.list_name IN ({})'.format(q))
             params.extend(lists)
-        if urgent:
-            extra_where.append('AND priority = 9')
+        if priority:
+            extra_where.append('AND PRIORITY > 0 AND PRIORITY <= ?')
+            params.append('{}'.format(priority))
         if location:
             extra_where.append('AND location LIKE ?')
             params.append('%{}%'.format(location))